+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 
 - #102 Add Primary Referrer column in Doctor listings
 - #85 Allow client contact to list/add/edit Batches from its own Client
+- #102 Add Primary Referrer column in Doctor listings
+- #85 Allow client contact to list/add/edit Cases from its own Client
 - #83 Allow client contact to create and edit Doctors
 
 **Changed**